Refactor scheduled circulars Lambda
I'm preparing to add yet a third action to generate sitemap files.
To keep everything nice and neat, here are a few refactorings:

- Collect all of the actions into the actions/ directory.
- Factor out repeated S3 bucket config details into a common file.
- Collect the mapCirculars function and the CircularsAction type
  into one file to keep the Lambda entry point itself clean as
  we add more imports for the actions.
- Rename the mapCirculars function to forAllCirculars, because its
  behavior more closely resembles the .forEach() array method than
  the .map() method.
